+++ src/components/ui/AuthModal/SignUp/SignUp.tsx
@@ -3,12 +3,11 @@ import { useAppDispatch } from '../../../../app/hooks';
 import { Alert, Button, Stack, Typography } from '@mui/material';
 import { Formik, Form, Field } from 'formik';
 import { TextField } from 'formik-mui';
-import { PasswordInput } from '../../..';
+import { ConfirmDialog, FullScreenLoader, PasswordInput } from '../../..';
 import getValidationSchema from './validationSchema';
 import signUpSubmitHandler from './submitHandler';
 import { ISignUpForm } from '../../../../types';
 import { useState } from 'react';
-import FullScreenLoader from '../../../FullScreenLoader/FullScreenLoader';
 
 interface ISignUpProps {
     closeModal: () => void;
@@ -24,10 +23,17 @@ const initialValues: ISignUpForm = {
 };
 
 const SignUp = ({ closeModal, setSignInAsSelectedTab }: ISignUpProps) => {
-    const { t } = useTranslation('auth');
+    const { t } = useTranslation(['auth', 'signUpDiscardChangesModal']);
     const dispatch = useAppDispatch();
 
     const [loading, setLoading] = useState(false);
+    const [isCloseConfirmModalDisplayed, setIsCloseConfirmModalDisplayed] =
+        useState<boolean>(false);
+
+    const signUpCloseHandler = () => {
+        closeModal();
+        setSignInAsSelectedTab();
+    };
 
     return (
         <>
@@ -46,13 +52,13 @@ const SignUp = ({ closeModal, setSignInAsSelectedTab }: ISignUpProps) => {
                     setLoading(false);
                 }}
             >
-                {({ isSubmitting }) => (
+                {({ isSubmitting, dirty }) => (
                     <Form noValidate>
                         <Stack gap={4}>
                             <Field
                                 fullWidth
                                 component={TextField}
-                                label={t('fullName')}
+                                label={t('auth:fullName')}
                                 name="fullName"
                                 required
                             />
@@ -60,7 +66,7 @@ const SignUp = ({ closeModal, setSignInAsSelectedTab }: ISignUpProps) => {
                             <Field
                                 fullWidth
                                 component={TextField}
-                                label={t('email')}
+                                label={t('auth:email')}
                                 type="email"
                                 name="email"
                                 required
@@ -69,7 +75,7 @@ const SignUp = ({ closeModal, setSignInAsSelectedTab }: ISignUpProps) => {
                             <Field
                                 fullWidth
                                 component={TextField}
-                                label={t('phone')}
+                                label={t('auth:phone')}
                                 name="phone"
                                 required
                             />
@@ -81,7 +87,7 @@ const SignUp = ({ closeModal, setSignInAsSelectedTab }: ISignUpProps) => {
                                 <Field
                                     fullWidth
                                     component={PasswordInput}
-                                    label={t('password')}
+                                    label={t('auth:password')}
                                     name="password"
                                     required
                                 />
@@ -89,7 +95,7 @@ const SignUp = ({ closeModal, setSignInAsSelectedTab }: ISignUpProps) => {
                                 <Field
                                     fullWidth
                                     component={PasswordInput}
-                                    label={t('confirmPassword')}
+                                    label={t('auth:confirmPassword')}
                                     name="confirmPassword"
                                     required
                                 />
@@ -97,7 +103,7 @@ const SignUp = ({ closeModal, setSignInAsSelectedTab }: ISignUpProps) => {
 
                             <Alert severity="info">
                                 <Typography component="p" variant="body2">
-                                    {t('passwordRequirements')}
+                                    {t('auth:passwordRequirements')}
                                 </Typography>
                             </Alert>
 
@@ -107,12 +113,20 @@ const SignUp = ({ closeModal, setSignInAsSelectedTab }: ISignUpProps) => {
                                 justifyContent={'center'}
                             >
                                 <Button
-                                    onClick={closeModal}
+                                    onClick={() => {
+                                        if (dirty) {
+                                            setIsCloseConfirmModalDisplayed(
+                                                true
+                                            );
+                                        } else {
+                                            signUpCloseHandler();
+                                        }
+                                    }}
                                     variant="outlined"
                                     fullWidth
                                     sx={{ textTransform: 'none' }}
                                 >
-                                    {t('cancel')}
+                                    {t('auth:cancel')}
                                 </Button>
                                 <Button
                                     type="submit"
@@ -121,7 +135,7 @@ const SignUp = ({ closeModal, setSignInAsSelectedTab }: ISignUpProps) => {
                                     disabled={isSubmitting}
                                     sx={{ textTransform: 'none' }}
                                 >
-                                    {t('signUp')}
+                                    {t('auth:signUp')}
                                 </Button>
                             </Stack>
                         </Stack>
@@ -129,6 +143,18 @@ const SignUp = ({ closeModal, setSignInAsSelectedTab }: ISignUpProps) => {
                 )}
             </Formik>
             <FullScreenLoader open={loading} />
+            <ConfirmDialog
+                title={t('signUpDiscardChangesModal:title')}
+                open={isCloseConfirmModalDisplayed}
+                handleClose={() => {
+                    setIsCloseConfirmModalDisplayed(false);
+                }}
+                agreeText={t('signUpDiscardChangesModal:confirmText')}
+                disagreeText={t('signUpDiscardChangesModal:cancelText')}
+                handleAgreeFunction={signUpCloseHandler}
+            >
+                {t('signUpDiscardChangesModal:message')}
+            </ConfirmDialog>
         </>
     );
 };