+++ src/layout/Header/Header.tsx
@@ -1,9 +1,10 @@
 import { useTranslation } from 'react-i18next';
 import { useState } from 'react';
 import { useTheme } from '@emotion/react';
+import { useNavigate } from 'react-router-dom';
 import { useAppDispatch, useAppSelector } from '../../app/hooks';
 import { toggleTheme } from '../../app/features/themeSlice';
-import { logout, showAuthModal } from '../../app/features/authSlice';
+import { showAuthModal } from '../../app/features/authSlice';
 import {
     Box,
     AppBar,
@@ -17,9 +18,7 @@ import {
     Divider,
     useMediaQuery,
 } from '@mui/material';
-import { toast } from 'react-toastify';
-import { useNavigate } from 'react-router-dom';
-import { ConfirmDialog } from '../../components';
+import { ConfirmDialog, FullScreenLoader } from '../../components';
 import { StyledProfileButton, StyledToolBar } from './Header.styled';
 import LightModeRoundedIcon from '@mui/icons-material/LightModeRounded';
 import DarkModeRoundedIcon from '@mui/icons-material/DarkModeRounded';
@@ -28,21 +27,27 @@ import KeyboardArrowDownIcon from '@mui/icons-material/KeyboardArrowDown';
 import PersonIcon from '@mui/icons-material/Person';
 import ExitToAppIcon from '@mui/icons-material/ExitToApp';
 import PermContactCalendarIcon from '@mui/icons-material/PermContactCalendar';
+import { BackupTable, PeopleAlt } from '@mui/icons-material';
 import { paths } from '../../config';
 import { EUserRole } from '../../types';
-import { BackupTable, PeopleAlt } from '@mui/icons-material';
+import { useLogout } from '../../hooks';
 
 export const Header = () => {
     const { t } = useTranslation(['headerFooter', 'logoutConfirmationModal']);
     const dispatch = useAppDispatch();
     const navigate = useNavigate();
     const theme = useTheme();
+    const logoutHandler = useLogout();
+
     const isSmallScreen = useMediaQuery(
         `(max-width:${theme.breakpointValues.small})`
     );
 
     const themeMode = useAppSelector((state) => state.theme.currentTheme);
     const user = useAppSelector((state) => state.auth.user);
+
+    const [logoutLoading, setLogoutLoading] = useState(false);
+
     const [isLogoutModalDisplayed, setIsLogoutModalDisplayed] =
         useState<boolean>(false);
 
@@ -64,10 +69,10 @@ export const Header = () => {
     };
 
     const handleLogoutClick = () => {
-        dispatch(logout());
+        setLogoutLoading(true);
+        void logoutHandler();
         handleCloseUserMenu();
-        toast.success(t('logoutConfirmationModal:logoutSuccessMessage'));
-        navigate(paths.home);
+        setLogoutLoading(false);
     };
 
     return (
@@ -255,6 +260,7 @@ export const Header = () => {
             >
                 {t('logoutConfirmationModal:message')}
             </ConfirmDialog>
+            <FullScreenLoader open={logoutLoading} />
         </>
     );
 };